* sdk: Fix bug #19025 'std::bad_alloc' when loading cmake generated project
When the wxFile is given a directory instead of a file the Length method
returns quite a big number, causing std::bad_allow exceptions. This is
cause by ill-formed cbp file, that has a <unit> tag that points to a
directory. Checking if the fileName points to a directory and exiting fixes
the crash.